refactored Challenge into extension files
