+++ Challenges/Challenge+CoreDataProperties.swift
@@ -10,7 +10,6 @@ import Foundation
 import CoreData
 import CoreGraphics
 
-
 extension Challenge {
 
     @nonobjc public class func fetchRequest() -> NSFetchRequest<Challenge> {
@@ -27,18 +26,6 @@ extension Challenge {
     @NSManaged public var sendReminders: Bool
     @NSManaged public var start: Date?
     @NSManaged public var actions: NSSet?
-    
-    var unwrappedName: String {
-        name ?? "Unknown"
-    }
-    
-    var actionsArray: [Action] {
-        let set = actions as? Set<Action> ?? []
-        return set.sorted {
-            $0.unwrappedDate < $1.unwrappedDate
-        }
-    }
-
 }
 
 // MARK: Generated accessors for actions
@@ -56,53 +43,4 @@ extension Challenge {
     @objc(removeActions:)
     @NSManaged public func removeFromActions(_ values: NSSet)
 
-}
-
-extension Challenge : Identifiable {
-
-}
-
-extension Challenge {
-    static var preview: Challenge {
-        let challenge = Challenge()
-        challenge.id = UUID()
-        challenge.start = Date()
-        challenge.name = "Challenge Preview"
-        challenge.frequency = "daily"
-        challenge.goal = 3000
-        challenge.regularGoal = 100
-        challenge.sendReminders = false
-        challenge.isActive = true
-        return challenge
-    }
-}
-
-extension Challenge {
-    
-    func dailyRepetitions(for date: Date) -> Double {
-        return actionsArray.filter { (action: Action) in
-            let formatter = DateFormatter()
-            formatter.dateFormat = "yyyy-MM-dd"
-            
-            return formatter.string(from: action.unwrappedDate) == formatter.string(from: date)
-        }.reduce(0.0) { (previousResult, action) -> Double in
-            previousResult + action.count
-        }
-    }
-    
-    var totalCount: Double {
-        return actionsArray
-            .map { $0.count }
-            .reduce(0.0, { accumulated, nextValue in
-                accumulated + nextValue
-            })
-    }
-    
-    func heightOfBar(for date: Date, with maxHeight: CGFloat) -> CGFloat {
-        let percentage = dailyRepetitions(for: date) / regularGoal
-        let goalHeight = maxHeight * 0.8
-        let result = goalHeight * CGFloat(percentage)
-        return result.clamped(to: 0...maxHeight)
-    }
-    
 }
